bump wasmtime to v27

+++ src/engine.rs
@@ -97,9 +97,9 @@ impl ResourceLimiter for MemoryLimiter {
     /// See [`wasmtime::ResourceLimiter::table_growing`].
     fn table_growing(
         &mut self,
-        _current: u32,
-        _desired: u32,
-        _maximum: Option<u32>,
+        _current: usize,
+        _desired: usize,
+        _maximum: Option<usize>,
     ) -> anyhow::Result<bool> {
         Ok(true)
     }
